Deep Learning Projects/Anime_Face_Generation_Using_GANs/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.2
matplotlib==3.8.2
Pillow==10.1.0
tqdm==4.66.1